Simple split of UI into separate classes
There is now MusicView, Tabs, and Timeline.  This is a very simple
split, with much nicer factorization possible as time goes on.
